:truck: rename: 챗 서비스 패키지 구조 변경
