+++ src/domain/chat/chat.module.ts
@@ -1,15 +1,17 @@
 import { Module } from '@nestjs/common';
-import { ChatService } from './chat.service';
+import { ChatService } from './service/chat.service';
 import { OpenAIModule } from '../../infrastructure/openai/openai.module';
-import {ChatDataService} from "./chat.data.servivce";
+import {ChatDataService} from "./service/chat.data.servivce";
 import {MongooseModule} from "@nestjs/mongoose";
 import {Post, PostSchema} from "../post/post.schema";
 import {ChatMessage, ChatMessageSchema} from "./schema/chat.message.schema";
 import {ChatSession, ChatSessionSchema} from "./schema/chat.session.schema";
+import {GeneratedContent, GeneratedContentSchema} from "./schema/generated.content.schema";
 
 @Module({
     imports: [MongooseModule.forFeature([{ name: ChatMessage.name, schema: ChatMessageSchema }
-        , {name: ChatSession.name, schema: ChatSessionSchema}])
+        , {name: ChatSession.name, schema: ChatSessionSchema}
+        , {name: GeneratedContent.name, schema: GeneratedContentSchema}])
         , OpenAIModule],
     providers: [ChatService, ChatDataService],
     exports: [ChatService, ChatDataService],
