+++ src/domain/kakao/kakao.module.ts
@@ -7,7 +7,7 @@ import {ArticleQueueModule} from "./queue/article-queue.module";
 import {BullModule} from "@nestjs/bull";
 import {UserModule} from "../user/user.module";
 import {PostModule} from "../post/post.module";
-import {ChatDataService} from "../chat/chat.data.servivce";
+import {ChatDataService} from "../chat/service/chat.data.servivce";
 
 @Module({
     imports: [ChatModule, HttpModule, UserModule, PostModule, ChatModule,
